// src/Platform/sdk.typescript/src/model/accessPointCreate.ts
/**
 * 2S-API
 * The 2S-API is a RESTful API for interacting with the 2S Platform.
 *
 * The version of the OpenAPI document: 0.8.26.4237
 * Contact: kobr.ales@outlook.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';

export class AccessPointCreate {
    'organization'?: string;
    'project'?: string;
    'ownerId'?: string;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "organization",
            "baseName": "Organization",
            "type": "string"
        },
        {
            "name": "project",
            "baseName": "Project",
            "type": "string"
        },
        {
            "name": "ownerId",
            "baseName": "OwnerId",
            "type": "string"
        }    ];

    static getAttributeTypeMap() {
        return AccessPointCreate.attributeTypeMap;
    }
}