// src/Platform/sdk.typescript/src/model/account.ts
/**
 * 2S-API
 * The 2S-API is a RESTful API for interacting with the 2S Platform.
 *
 * The version of the OpenAPI document: 0.8.26.4237
 * Contact: kobr.ales@outlook.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';

export class Account {
    'id'?: string;
    'userName'?: string;
    'name'?: string;
    'accessMap'?: { [key: string]: string; };

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "id",
            "baseName": "Id",
            "type": "string"
        },
        {
            "name": "userName",
            "baseName": "UserName",
            "type": "string"
        },
        {
            "name": "name",
            "baseName": "Name",
            "type": "string"
        },
        {
            "name": "accessMap",
            "baseName": "AccessMap",
            "type": "{ [key: string]: string; }"
        }    ];

    static getAttributeTypeMap() {
        return Account.attributeTypeMap;
    }
}

export namespace Account {
    export enum AccessMapEnum {
        None = <any> 'None',
        Reader = <any> 'Reader',
        Contributor = <any> 'Contributor',
        Administrator = <any> 'Administrator',
        Owner = <any> 'Owner'
    }
}