// src/Platform/sdk.typescript/src/model/errorResponse.ts
/**
 * 2S-API
 * The 2S-API is a RESTful API for interacting with the 2S Platform.
 *
 * The version of the OpenAPI document: 0.8.26.4237
 * Contact: kobr.ales@outlook.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';
import { ExceptionWrapper } from './exceptionWrapper';

export class ErrorResponse {
    'message'?: string;
    'hint'?: string;
    'errorCode'?: string;
    'error'?: ExceptionWrapper;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "message",
            "baseName": "Message",
            "type": "string"
        },
        {
            "name": "hint",
            "baseName": "Hint",
            "type": "string"
        },
        {
            "name": "errorCode",
            "baseName": "ErrorCode",
            "type": "string"
        },
        {
            "name": "error",
            "baseName": "Error",
            "type": "ExceptionWrapper"
        }    ];

    static getAttributeTypeMap() {
        return ErrorResponse.attributeTypeMap;
    }
}